requirements.txt
langchain>=0.1.0
langchain-community>=0.0.10
sentence-transformers>=2.2.2
torch>=2.0.0
transformers>=4.30.0
pypdf>=3.0.0
python-dotenv==1.0.0
ipywidgets>=8.0.0
tqdm>=4.65.0
numpy>=1.21.0
pandas>=1.3.0
jupyter>=1.0.0
notebook>=6.4.0
google-generativeai>=0.3.0
pinecone-client>=3.0.0
openai>=1.0.0
requests>=2.28.0 